Extract `safe_args` to a jvm backend module.
The utility was only used by jvm tasks.
