+++ src/python/pants/backend/jvm/tasks/BUILD
@@ -275,6 +275,7 @@ python_library(
     '3rdparty/python:six',
     ':classpath_util',
     ':nailgun_task',
+    'src/python/pants/backend/jvm:argfile',
     'src/python/pants/backend/jvm/subsystems:jar_tool',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
@@ -333,25 +334,28 @@ python_library(
   sources = ['junit_run.py'],
   dependencies = [
     '3rdparty/python:six',
+    '3rdparty/python/twitter/commons:twitter.common.collections',
     ':classpath_util',
     ':coverage',
     ':jvm_task',
     ':jvm_tool_task_mixin',
     ':reports',
+    'src/python/pants/backend/jvm:argfile',
     'src/python/pants/backend/jvm/subsystems:jvm_platform',
     'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:build_environment',
+    'src/python/pants/base:exceptions',
     'src/python/pants/base:workunit',
-    'src/python/pants/binaries:binary_util',
     'src/python/pants/build_graph',
-    'src/python/pants/java:util',
+    'src/python/pants/java/distribution',
+    'src/python/pants/java:executor',
     'src/python/pants/task',
     'src/python/pants/util:argutil',
     'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-    'src/python/pants/util:process_handler',
+    'src/python/pants/util:desktop',
+    'src/python/pants/util:strutil',
     'src/python/pants/util:xml_parser',
   ],
 )
